Rename the FootNote component
In addition, uses a h2 tag as its root element, since it is more of a title than a foot note.
