+++ components/section-title/index.js
