+++ components/section-title/style.scss
@@ -1,23 +1,25 @@
 @import 'css/settings';
 
-.c-foot-note {
+.c-section-title {
   display: flex;
   flex-direction: column;
   align-items: center;
   color: $white;
   font-family: $font-family-3;
   font-weight: $font-weight-medium;
-  font-size: rem(12);
+  font-size: $font-size-sm;
   letter-spacing: rem(3);
+  text-transform: uppercase;
   line-height: rem(28);
   padding: rem(20);
   width: 100%;
 
-  .page-number {
+  .index {
     box-sizing: border-box;
     padding: 0 rem(5);
     width: rem(30);
     height: rem(30);
+    margin-bottom: rem(15);
     border: rem(2) solid $white;
     height: 100%;
   }
